+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # class-factory-phpstorm Changelog
 
 ## [Unreleased]
+### Fixed
+- Incorrect reports of unknown columns when using nested states
 
 ## [2.0.1]
 ### Fixed