+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # class-factory-phpstorm Changelog
 
 ## [Unreleased]
+## [2.0.2]
 ### Fixed
 - Incorrect reports of unknown columns when using nested states
 - Class property inspection when using nested states